Removed a cast that is not needed now.
git-svn-id: https://svn.opendtect.org/od/trunk@38556 a985855d-d384-4b45-ae33-31a410998c3c

+++ src/Seis/seisdatapackzaxistransformer.cc
@@ -91,8 +91,7 @@ bool SeisDataPackZAxisTransformer::doWork(
     const StepInterval<float>& inpzrg = seisdp->getZRange();
     const int nrinpsamp = inpzrg.nrSteps() + 1;
     const int nroutsamp = zrange_.nrSteps() + 1;
-    const int nrtrcs = (int)(seisdp->nrTrcs()
-			  / seisdp->data(0).info().getSize(0));
+    const int nrtrcs = seisdp->nrTrcs() / seisdp->data(0).info().getSize(0);
 
     ZAxisTransformSampler outputsampler( transform_, true,
 	    SamplingData<double>(zrange_.start, zrange_.step), false );
